requirements.txt
# Core LangChain ecosystem
langchain>=0.2.14
langchain-core>=0.2.30
langchain-community>=0.2.12
langchain-google-genai>=1.0.8

# LCEL / Runnable support (comes with core but keeping explicit)
langchainhub>=0.1.20

# Tavily Search Tool
tavily-python>=0.3.3

# Web Scraping
beautifulsoup4>=4.12.3
requests>=2.32.3
lxml>=5.2.2

# Environment variables 
python-dotenv>=1.0.1

# Async support & HTTP clients
httpx>=0.27.0
aiohttp>=3.9.5

# Parsing / utilities
pydantic>=2.7.4
tqdm>=4.66.4

# Optional (but highly recommended for research systems)
# Vector store (if you later extend to RAG)
chromadb>=0.5.5

# Token handling / text processing
tiktoken>=0.7.0

# Logging & debugging
rich>=13.7.1

# Jupyter (optional for testing)
notebook>=7.2.1
ipykernel>=6.29.5

# FastAPI (if you expose this as API later)
fastapi>=0.111.0
uvicorn>=0.30.1


# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── LangChain ─────────────────────────────────────────────────────────────────
langchain==0.3.25
langchain-core==0.3.52
langchain-groq==0.3.2
langgraph==0.4.1

# ── LLM / AI ──────────────────────────────────────────────────────────────────
groq==0.28.0

# ── Search & Scraping ─────────────────────────────────────────────────────────
tavily-python==0.5.1
requests==2.32.3
beautifulsoup4==4.12.3

# ── Data / Utilities ──────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.4

# ── Serper image search uses requests (already above) ─────────────────────────
# No extra package needed — Serper API is called via requests.post()